+++ transportnetwork.h
@@ -51,7 +51,7 @@ public:
     bool exists(const std::string& code) const;
     double dijkstraDistance(const std::string& code1, const std::string& code2); 
     std::list<std::string> dijkstraPath(const std::string& code1, const std::string& code2); 
-    int bfsDistance(const std::string& code1, const std::string& code2);
+    double bfsDistance(const std::string& code1, const std::string& code2);
     std::list<std::string> bfsPath(const std::string& code1, const std::string& code2);
 };
